Update photo editor module
